feature: refactor TsmlUi.tsx & add StateContext
- Seperated out global state from `TsmlUi.tsx` to `StateContext.tsx`
- Routing mostly works, there is an issue in Table.tsx with the onClick
that needs to be fixed
