+++ src/components/TsmlUI.tsx
@@ -1,243 +1,70 @@
-import { useEffect, useState } from 'react';
+import { useContext, useEffect } from 'react';
 import { Global } from '@emotion/react';
+import { IStateContext } from './StateContext';
 
 import { globalCss } from '../styles';
-import type { State } from '../types';
 import { Alert, Controls, Loading, Map, Meeting, Table, Title } from './';
 
-import {
-  filterMeetingData,
-  getQueryString,
-  loadMeetingData,
-  mergeSettings,
-  setQueryString,
-  SettingsContext,
-  translateGoogleSheet,
-} from '../helpers';
+import { setQueryString, SettingsContext, getQueryString } from '../helpers';
 import { useLocation, useNavigate } from 'react-router-dom';
+import { StateContext } from './StateContext';
+import { formatUrl } from '../helpers';
 
-export default function TsmlUI({
-  google,
-  mapbox,
-  settings: userSettings,
-  src,
-  timezone,
-}: {
-  google?: string;
-  mapbox?: string;
-  settings?: TSMLReactConfig;
-  src?: string;
-  timezone?: string;
-}) {
-  const [state, setState] = useState<State>({
-    capabilities: {
-      coordinates: false,
-      distance: false,
-      geolocation: false,
-      inactive: false,
-      location: false,
-      region: false,
-      sharing: false,
-      time: false,
-      type: false,
-      weekday: false,
-    },
-    input: {
-      distance: [],
-      mode: 'search',
-      region: [],
-      time: [],
-      type: [],
-      view: 'table',
-      weekday: [],
-    },
-    indexes: {
-      distance: [],
-      region: [],
-      time: [],
-      type: [],
-      weekday: [],
-    },
-    loading: true,
-    meetings: {},
-    ready: false,
-  });
-
-  const { settings, strings } = mergeSettings(userSettings);
-
+export default function TsmlUI() {
+  const {
+    state,
+    setState,
+    settings,
+    mapbox,
+    strings,
+    inProgress,
+    filteredSlugs,
+    loadData,
+  } = useContext(StateContext) as IStateContext;
   const location = useLocation();
   const navigate = useNavigate();
-
-  // enable forward & back buttons
-  // useEffect(() => {
-  //   const popstateListener = () => {
-  //     setState({ ...state, input: getQueryString(settings, location) });
-  //   };
-  //   window.addEventListener('popstate', popstateListener);
-
-  //   // update canonical
-  //   let canonical = document.querySelector('link[rel="canonical"]');
-  //   if (!canonical) {
-  //     canonical = document.createElement('link');
-  //     canonical.setAttribute('rel', 'canonical');
-  //     document.getElementsByTagName('head')[0]?.appendChild(canonical);
-  //   }
-  //   canonical.setAttribute(
-  //     'href',
-  //     formatUrl(
-  //       state.input.meeting ? { meeting: state.input.meeting } : state.input,
-  //       settings
-  //     )
-  //   );
-
-  //   return () => {
-  //     window.removeEventListener('popstate', popstateListener);
-  //   };
-  // }, [state, location]);
 
   // manage classes
   useEffect(() => {
     document.body.classList.add('tsml-ui');
+    loadData();
     return () => {
       document.body.classList.remove('tsml-ui');
     };
   }, []);
 
-  // load data once
-  if (state.loading) {
-    console.log(
-      'TSML UI meeting finder: https://github.com/code4recovery/tsml-ui'
-    );
-
-    const input = getQueryString(settings, location);
-
-    if (!src) {
-      setState({
-        ...state,
-        error: 'Configuration error: a data source must be specified.',
-        loading: false,
-        ready: true,
-      });
-    } else {
-      const sheetId = src.startsWith('https://docs.google.com/spreadsheets/d/')
-        ? src.split('/')[5]
-        : undefined;
-
-      // google sheet
-      if (sheetId) {
-        if (!google) {
-          setState({
-            ...state,
-            error: 'Configuration error: a Google API key is required.',
-            loading: false,
-          });
-        }
-        src = `https://sheets.googleapis.com/v4/spreadsheets/${sheetId}/values/A1:ZZ?key=${google}`;
-      }
-
-      // cache busting
-      if (src.endsWith('.json') && input.meeting) {
-        src = `${src}?${new Date().getTime()}`;
-      }
-
-      // fetch json data file and build indexes
-      fetch(src)
-        .then(res => (res.ok ? res.json() : Promise.reject(res.status)))
-        .then(data => {
-          if (sheetId) {
-            data = translateGoogleSheet(data, sheetId, settings);
-          }
-
-          if (!Array.isArray(data) || !data.length) {
-            return setState({
-              ...state,
-              error: 'Configuration error: data is not in the correct format.',
-              loading: false,
-              ready: true,
-            });
-          }
-
-          const [meetings, indexes, capabilities] = loadMeetingData(
-            data,
-            state.capabilities,
-            settings,
-            strings,
-            timezone
-          );
-
-          if (!timezone && !Object.keys(meetings).length) {
-            return setState({
-              ...state,
-              error: 'Configuration error: time zone is not set.',
-              loading: false,
-              ready: true,
-            });
-          }
-
-          const waitingForGeo =
-            (!input.latitude || !input.longitude) &&
-            ((input.mode === 'location' && input.search) ||
-              input.mode === 'me');
-
-          setState({
-            ...state,
-            capabilities,
-            indexes,
-            input,
-            loading: false,
-            meetings,
-            ready: !waitingForGeo,
-          });
-        })
-        .catch(error => {
-          const errors = {
-            400: 'bad request',
-            401: 'unauthorized',
-            403: 'forbidden',
-            404: 'not found',
-            429: 'too many requests',
-            500: 'internal server',
-            502: 'bad gateway',
-            503: 'service unavailable',
-            504: 'gateway timeout',
-          };
-          setState({
-            ...state,
-            error: errors[error as keyof typeof errors]
-              ? `Error: ${
-                  errors[error as keyof typeof errors]
-                } (${error}) when ${
-                  sheetId ? 'contacting Google' : 'loading data'
-                }.`
-              : error.toString(),
-            loading: false,
-            ready: true,
-          });
-        });
+  useEffect(() => {
+    // apply input changes to query string
+    if (!state.loading) {
+      setQueryString(state.input, settings!, navigate, location);
     }
-  }
+  }, [state.loading]);
 
-  // apply input changes to query string
-  if (!state.loading) {
-    setQueryString(state.input, settings, navigate, location);
-  }
+  useEffect(() => {
+    // if location has changed, update state
+    if (state.input !== getQueryString(settings, location)) {
+      setState({ ...state, input: getQueryString(settings, location) });
+    }
 
-  // filter data
-  const [filteredSlugs, inProgress] = filterMeetingData(
-    state,
-    setState,
-    settings,
-    strings,
-    mapbox
-  );
+    let canonical = document.querySelector('link[rel="canonical"]');
+    if (!canonical) {
+      canonical = document.createElement('link');
+      canonical.setAttribute('rel', 'canonical');
+      document.getElementsByTagName('head')[0]?.appendChild(canonical);
+    }
+    canonical.setAttribute(
+      'href',
+      formatUrl(
+        state.input.meeting ? { meeting: state.input.meeting } : state.input,
+        settings
+      )
+    );
+  }, [location]);
 
-  // show alert?
-  state.alert = !filteredSlugs.length ? strings.no_results : undefined;
-
-  // show error?
-  if (state.input.meeting && !state.meetings[state.input.meeting]) {
-    state.error = strings.not_found;
-  }
+  // useEffect(() => {
+  //   // if state has changed update location
+  //   navigate(formatRelativeUrl(state.input, settings, location));
+  // }, [state]);
 
   return (
     <SettingsContext.Provider value={{ settings, strings }}>
